pyproject.toml
[project]
name = "scholarwriting"
version = "0.1.0"
description = "Multi-agent academic writing framework with platform-neutral workflow contracts."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "jsonschema>=4.0",
    "pyyaml>=6.0",
]

[project.scripts]
scholar-writing = "scholar_writing.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scholar_writing"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
